update(files): monorepo with backend and front end src; package.json
