+++ frontend/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "eduhope-frontend",
+  "version": "0.1.0",
+  "private": true,
+  "description": "EduHope.SG website front end",
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
+  },
+  "dependencies": {
+    "next": "^12.3.1",
+    "react": "17.0.1",
+    "react-dom": "17.0.1"
+  },
+  "author": "EduHope Team",
+  "license": "agpl-3.0"
+}
