+++ frontend/pages/hello.js
@@ -1,0 +1,4 @@
+export default function Hello() {
+    return <h1>EDUHOPE</h1>
+  }
+
