+++ frontend/public/CHEMISTRY(H2).json
@@ -1,0 +1,9 @@
+{
+    "id": "CHEMISTRY(H2)",
+    "color": "orange",
+    "image": "https://images.unsplash.com/photo-1519440862171-af26cf8c2a85?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1352&q=80",
+    "Other": "Chemistry, Mathematics",
+    "Medium": "Text, virtual consults",
+    "Education": " Secondary 4",
+    "Description": "I am able to do crash courses as well as"
+}
