+++ frontend/public/ENGLISH.json
@@ -1,0 +1,9 @@
+{
+    "id": "ENGLISH",
+    "color": "blue",
+    "image": "https://images.unsplash.com/photo-1551830820-330a71b99659?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1350&q=80",
+    "Other": "Chemistry, Mathematics",
+    "Medium": "Text, virtual consults",
+    "Education": " Secondary 4",
+    "Description": "I am able to do crash courses as well as"
+}
