+++ frontend/public/MATHEMATICS(H2).json
@@ -1,0 +1,9 @@
+{
+    "id": "MATHEMATICS(H2)",
+    "color": "white",
+    "image": "https://images.unsplash.com/photo-1561580125-028ee3bd62eb?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1350&q=80",
+    "Other": "Chemistry, Mathematics",
+    "Medium": "Text, virtual consults",
+    "Education": " Secondary 4",
+    "Description": "I am able to do crash courses as well as other stuff"
+}
