--- package.json
@@ -1,15 +1,0 @@
-{
-  "name": "next-basic",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "next": "^12.3.1",
-    "react": "17.0.1",
-    "react-dom": "17.0.1"
-  }
-}
