--- pages/hello.js
@@ -1,4 +1,0 @@
-export default function Hello() {
-    return <h1>EDUHOPE</h1>
-  }
-
